#195 Address review feedback: extract setup constants and hoist helpers
Extract SetupStep type, SetupStepperState interface, polling constants, and
statusToStep helper from use-setup-stepper.ts into a new
src/features/vehicles/lib/setup-constants.ts module, reducing the hook file
from 257 to 219 lines. Hoist SETUP_STATUS_ORDER and advanceSetupStatus from
inside the syncVehiclesFromTesla for-loop to module scope in sync.ts.
