+++ src/features/vehicles/hooks/use-setup-stepper.ts
@@ -2,62 +2,25 @@
 
 import { useState, useEffect, useRef, useCallback } from 'react';
 
-import type { SetupStatus } from '@/types/vehicle';
 import { checkPairingStatus, checkVehicleConnection, updateSetupStatus } from '../api/setup-actions';
 import { pushFleetConfig } from '../api/fleet-config';
-
-// ─── Constants ────────────────────────────────────────────────────────────────
-
-/** Exponential backoff settings for Step 1 pairing poll. */
-const PAIRING_POLL_BASE_MS = 3_000;
-const PAIRING_POLL_MAX_MS = 30_000;
-const PAIRING_POLL_MULTIPLIER = 1.5;
-
-/** Fixed interval for Step 3 connection poll (every 5s). */
-const CONNECTION_POLL_INTERVAL_MS = 5_000;
-
-/** Delay (ms) before showing the "vehicle may be offline" helper text. */
-const OFFLINE_HELPER_DELAY_MS = 300_000; // 5 minutes
-
-/** How long to show Step 4 (Connected) before auto-dismissing. */
-const CONNECTED_AUTO_DISMISS_MS = 3_000;
-
-// ─── Types ────────────────────────────────────────────────────────────────────
-
-export type SetupStep = 1 | 2 | 3 | 4;
-
-export interface SetupStepperState {
-  /** Current step (1-4). */
-  step: SetupStep;
-  /** Whether the current step is polling/loading. */
-  isLoading: boolean;
-  /** Error message, if the current step has errored. */
-  error: string | null;
-  /** True after 5 min of waiting in Step 3 (vehicle may be offline). */
-  showOfflineHelper: boolean;
-  /** True when setup is complete and the banner should be dismissed. */
-  isDismissed: boolean;
-  /** Retry the current failed step. */
-  onRetry: () => void;
-}
+import type { SetupStatus } from '@/types/vehicle';
+import {
+  PAIRING_POLL_BASE_MS,
+  PAIRING_POLL_MAX_MS,
+  PAIRING_POLL_MULTIPLIER,
+  CONNECTION_POLL_INTERVAL_MS,
+  OFFLINE_HELPER_DELAY_MS,
+  CONNECTED_AUTO_DISMISS_MS,
+  statusToStep,
+} from '../lib/setup-constants';
+export type { SetupStep, SetupStepperState } from '../lib/setup-constants';
 
 export interface UseSetupStepperOptions {
   vehicleId: string;
   vin: string;
   userId: string;
   initialSetupStatus: SetupStatus;
-}
-
-// ─── Status → Step mapping ────────────────────────────────────────────────────
-
-function statusToStep(status: SetupStatus): SetupStep {
-  switch (status) {
-    case 'pending_pairing': return 1;
-    case 'pairing_detected': return 2;
-    case 'config_pushed': return 3;
-    case 'waiting_connection': return 3;
-    case 'connected': return 4;
-  }
 }
 
 // ─── Hook ─────────────────────────────────────────────────────────────────────
@@ -75,8 +38,8 @@ export function useSetupStepper({
   vin,
   userId,
   initialSetupStatus,
-}: UseSetupStepperOptions): SetupStepperState {
-  const [step, setStep] = useState<SetupStep>(statusToStep(initialSetupStatus));
+}: UseSetupStepperOptions) {
+  const [step, setStep] = useState(statusToStep(initialSetupStatus));
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [showOfflineHelper, setShowOfflineHelper] = useState(false);
